Performs and Approximate Bayesian Computation Sampling of Model Parameters

Given a set of simulation experiments (list), a model, parameter boundaries, this function will draw a sample of parameters from the posterior probability density of the given problem.

Usage

ABCMCMC(
  objectiveFunction,
  startPar,
  nSims,
  Sigma0,
  delta,
  dprior,
  batchSize = 100
)

Arguments

objectiveFunction

function that, given a (vectorial) parameter as input, simulates the model, and outputs the distance between experimental data and data simulated from the model with the parameter provided in input

startPar

starting value for the parameter vector

nSims

requested sample size

Sigma0

multivariate normal covariance of Markov chain transition kernel

delta

ABC acceptance threshold

dprior

a function that returns prior probability density values

Value

a list containing a sample matrix and a vector of scores (values of delta for each sample)

Details

Initially this function performs a similar job as an optimizer, and then transitions to MCMC sampling.