rCopulaPrior returns a function that generates random values from the copula model

The returned function generates n random vectors, as rows of a matrix.

Usage

rCopulaPrior(Copula)

Arguments

Copula

the return value of fitCopula()

Value

a matrix of random values

Examples

rprior <- rNormalPrior(c(-1,0,1),c(1,2,3))
C <- fitCopula(rprior(1000))
D <- rCopulaPrior(C)
print(cov(D(100)))
#>             [,1]        [,2]       [,3]
#> [1,]  1.02389567 -0.07665333 -0.3028752
#> [2,] -0.07665333  4.42906472  0.4791655
#> [3,] -0.30287520  0.47916549  9.1234128
print(D(10))
#>             [,1]        [,2]       [,3]
#>  [1,] -1.1976316 -1.70409997 -1.8511321
#>  [2,] -1.3736909 -1.98311580 -2.0798774
#>  [3,] -0.7211491  1.66289672 -0.8608076
#>  [4,] -0.7234431  0.70693355 -1.2941368
#>  [5,]  0.3170828 -2.32966016  7.5167396
#>  [6,] -1.5987839  1.02096560  4.7890459
#>  [7,] -1.8243182  1.66384531  3.3652893
#>  [8,] -0.9733366  2.08247707 -2.8356582
#>  [9,] -1.1985233  0.06072809 -3.2707875
#> [10,] -0.9595520 -2.22821360  1.0877956