rCopulaPrior returns a function that generates random values from the copula model

The returned function generates n random vectors, as rows of a matrix.

Usage

rCopulaPrior(Copula)

Arguments

Copula

the return value of fitCopula()

Value

a matrix of random values

Examples

rprior <- rNormalPrior(c(-1,0,1),c(1,2,3))
C <- fitCopula(rprior(1000))
D <- rCopulaPrior(C)
print(cov(D(100)))
#>            [,1]        [,2]       [,3]
#> [1,] 1.16145963  0.08178159  0.1444280
#> [2,] 0.08178159  5.18566544 -0.6429784
#> [3,] 0.14442800 -0.64297842  9.7337067
print(D(10))
#>              [,1]       [,2]       [,3]
#>  [1,] -1.52086787 -1.4127466 -1.7880894
#>  [2,] -0.83326737 -1.3404273  4.0198509
#>  [3,]  0.05271797 -2.6468703 -0.5146367
#>  [4,] -0.78676684 -3.4976053  1.9293452
#>  [5,] -1.49000743  0.4214121 -2.1761898
#>  [6,]  0.10846422  2.2931669  7.3178558
#>  [7,] -0.86276814 -2.6004241 -5.9344651
#>  [8,] -1.30927322 -0.4875232  3.8963337
#>  [9,] -0.92589973 -0.3798322  4.7778896
#> [10,] -1.49428359  1.1568083  2.5500913