rCopulaPrior returns a function that generates random values from the copula model

The returned function generates n random vectors, as rows of a matrix.

Usage

rCopulaPrior(Copula)

Arguments

Copula

the return value of fitCopula()

Value

a matrix of random values

Examples

rcp<-rCopulaPrior(makeIndepCopula(ll=c(0,1,2),ul=c(1,2,3)))
rcp(12)
#>             [,1]     [,2]     [,3]
#>  [1,] 0.71570670 1.435282 2.239906
#>  [2,] 0.78472137 1.512122 2.764725
#>  [3,] 0.41556652 1.848060 2.821770
#>  [4,] 0.08141467 1.431830 2.338384
#>  [5,] 0.23810628 1.973249 2.275977
#>  [6,] 0.06730051 1.275485 2.588850
#>  [7,] 0.73975403 1.856392 2.824630
#>  [8,] 0.71453040 1.764069 2.180498
#>  [9,] 0.40745450 1.725221 2.345026
#> [10,] 0.58925087 1.660620 2.214232
#> [11,] 0.14387197 1.659785 2.380916
#> [12,] 0.15943314 1.886527 2.294505