rCopulaPrior returns a function that generates random values from the copula model

The returned function generates n random vectors, as rows of a matrix.

Usage

rCopulaPrior(Copula)

Arguments

Copula

the return value of fitCopula()

Value

a matrix of random values

Examples

rcp<-rCopulaPrior(makeIndepCopula(ll=c(0,1,2),ul=c(1,2,3)))
rcp(12)
#>             [,1]     [,2]     [,3]
#>  [1,] 0.01629761 1.801383 2.877017
#>  [2,] 0.52577326 1.972425 2.265952
#>  [3,] 0.34841997 1.270071 2.650221
#>  [4,] 0.87191439 1.918566 2.965500
#>  [5,] 0.28187890 1.640095 2.337046
#>  [6,] 0.94592205 1.009281 2.482915
#>  [7,] 0.93559600 1.685956 2.072764
#>  [8,] 0.35247819 1.322961 2.713175
#>  [9,] 0.70279259 1.785625 2.932886
#> [10,] 0.42456914 1.265795 2.803648
#> [11,] 0.32576995 1.642336 2.316926
#> [12,] 0.10211722 1.863657 2.395305