rCopulaPrior returns a function that generates random values from the copula model

The returned function generates n random vectors, as rows of a matrix.

Usage

rCopulaPrior(Copula)

Arguments

Copula

the return value of fitCopula()

Value

a matrix of random values

Examples

rcp<-rCopulaPrior(makeIndepCopula(ll=c(0,1,2),ul=c(1,2,3)))
rcp(12)
#>             [,1]     [,2]     [,3]
<<<<<<< HEAD
#>  [1,] 0.38519357 1.911539 2.821962
#>  [2,] 0.75879942 1.459000 2.492440
#>  [3,] 0.57571679 1.308567 2.872539
#>  [4,] 0.16738764 1.007963 2.797454
#>  [5,] 0.39146733 1.833847 2.640879
#>  [6,] 0.06197012 1.095433 2.575209
#>  [7,] 0.88619032 1.549634 2.399600
#>  [8,] 0.82733215 1.892302 2.715134
#>  [9,] 0.11224285 1.330425 2.859161
#> [10,] 0.16823977 1.246998 2.327707
#> [11,] 0.10317282 1.987275 2.479493
#> [12,] 0.29525650 1.899071 2.878998
=======
#>  [1,] 0.26553270 1.960789 2.007406
#>  [2,] 0.82222899 1.917377 2.322048
#>  [3,] 0.71683308 1.392364 2.675933
#>  [4,] 0.17615495 1.046590 2.182826
#>  [5,] 0.64326749 1.114925 2.705994
#>  [6,] 0.90997259 1.973176 2.290182
#>  [7,] 0.83250152 1.164213 2.908934
#>  [8,] 0.77951924 1.980218 2.971553
#>  [9,] 0.01219415 1.453190 2.541245
#> [10,] 0.17467041 1.846747 2.010251
#> [11,] 0.30967476 1.385138 2.847509
#> [12,] 0.07744600 1.339983 2.054661
>>>>>>> 55796ed (Added articles)