rUniformPrior returns a random vector generator

The return value is a function that generates random vectors of the same size as ll and ul from a uniform distribution within the limits defined by ul and ll. The random vectors are returned as n rows of a matrix, where n is the only argument of the returned function.

Usage

rUniformPrior(ll, ul)

Arguments

ll

lower limit of the random variables (a vector)

ul

upper limit of the random variables (same size vector as ll)

Value

a uniform random vector generating function: runiform(n), where n is the requested number of vectors (rows)

Examples

rup<-rUniformPrior(ll=c(0,1,2),ul=c(1,2,3))
rup(12)
#>             [,1]     [,2]     [,3]
#>  [1,] 0.63800028 1.131954 2.523796
#>  [2,] 0.01546784 1.730922 2.589015
#>  [3,] 0.89735055 1.557863 2.095538
#>  [4,] 0.76970891 1.802389 2.938171
#>  [5,] 0.66055086 1.628115 2.840644
#>  [6,] 0.33395491 1.391518 2.333552
#>  [7,] 0.03052661 1.564281 2.937988
#>  [8,] 0.57094218 1.666221 2.536167
#>  [9,] 0.30518194 1.313868 2.902429
#> [10,] 0.21842157 1.463625 2.444530
#> [11,] 0.90940322 1.527498 2.379092
#> [12,] 0.58049391 1.488352 2.941712