This creates a closure that simulates the model

This is a shorter alternative to the runModel function (C backend).

Usage

simulator.c(experiments, modelName, parMap = identity, noise = FALSE)

Arguments

experiments

a list of experiments to simulate: inital values, inputs, time vectors, initial times

modelName

a string (with optional comment indicating an .so file) which points out the model to simulate

parMap

the model will be called with parMap(parABC); so any parameter transformation can happen there.

noise

boolean variable. If noise=TRUE, Gaussian noise is added to the output of the simulations. The standard deviation of the Gaussian noise is equal to the measurement error. If noise=FALSE the output is the deterministic solution of the ODE system.

parABC

the parameters for the model, subject to change by parMap.

Value

a closure that returns the model's output for a given parameter vector

Details

It returns a closure around: - experiments, - the model, and - parameter mapping

The returned function depends only on parABC (the sampling parameters). The simulation will be done suing the rgsl backend.

Examples

   model.sbtab <- SBtabVFGEN::sbtab_from_tsv(dir(pattern="[.]tsv$"))
#> Error in file(con, "r"): invalid 'description' argument
   experiments <- SBtabVFGEN::sbtab.data(model.sbtab)
#> Error in is.factor(x): object 'model.sbtab' not found
   parABC <- SBtabVFGEN::sbtab.quantity(model.sbtab$Parameter)
#> Error: 'sbtab.quantity' is not an exported object from 'namespace:SBtabVFGEN'

   modelName <- checkModel("<insert_model_name>_gvf.c")
#> Error in checkModel("<insert_model_name>_gvf.c"): !is.null(modelFile) && length(modelFile) > 0 is not TRUE
   simulate <- simulator.c(experiments, modelName,  parABC)
#> Loading required package: rgsl
   yf <- sim(parABC)
#> Error in sim(parABC): could not find function "sim"