Load an example model for this package

This function finds the path to an example model, given by name. In the SBtab format, model and data travel together (in different tables, but the same documents).

Usage

uqsa_example(
  modelName = NULL,
  full.names = TRUE,
  pattern = "[.]tsv$",
  f = NULL
)

Arguments

modelName

name of model, e.g.: "AKAR4", "AKAP79", "CaMKII"; if empty, this function lists all available examples.

full.names

return full paths to files - defaults to TRUE

pattern

pattern to find specific files; if NULL, this function returns the directory of the example

f

file ending, search for file endings in f, alternative to pattern

Value

The location of the examples in the current environment if called with no arguments, the paths to the model files if a modelName was provided or the full path to the example if the file pattern pattern is unset

Details

By default this function returns the names of the tsv files belonging to the named model. If no modelName is provided it returns possible names (contents of the top-level example directory).

Examples

uqsa_example()
#> [1] "AKAP79"    "AKAP79tc"  "AKAR4"     "AKAR4cl"   "CaMKII"    "README.md"
uqsa_example("AKAR4",full.names=FALSE)
#> [1] "AKAR4_100nM.tsv"       "AKAR4_25nM.tsv"        "AKAR4_400nM.tsv"      
#> [4] "AKAR4_Compound.tsv"    "AKAR4_Experiments.tsv" "AKAR4_Output.tsv"     
#> [7] "AKAR4_Parameter.tsv"   "AKAR4_Reaction.tsv"   
uqsa_example("AKAP79",f='R',full.names=FALSE)
#> [1] "AKAP79.R"            "AKAP79_out.R"        "AKAP79_plots.R"     
#> [4] "configAKAP79.R"      "runABCMCMC_AKAP79.R"
uqsa_example("AKAP79",pat="^run.*R$")
#> [1] "/tmp/RtmpPHuiJ6/temp_libpath29167367bf1e3c/uqsa/extdata/AKAP79/runABCMCMC_AKAP79.R"